- Add PLY file loading. - Update some comments. - With meshes with many triangles, thousands of triangles can intersect multiple octree boxes. Previous implementation kept these faces in a parent box that completely contaned such triangles. This causes out-of-bounds errors during ray/mesh intersection using fixed maximum number of intersection face candidates. Instead, be greedy and put triangle into first child found that intersects triangle.

+++ GCRT/coshemisampler.cpp
@@ -44,5 +44,5 @@ uint32_t CosHemiSampler::NextSet()
 
     uint32_t outSet = curSet;
 
-    return (outSet >= samples.size() ? samples.size() - 1 : outSet);
+    return (uint32_t)(outSet >= samples.size() ? samples.size() - 1 : outSet);
 }
